Cleanup top-level controller
